--- components/RootLayout/RootLayout.module.scss
@@ -1,22 +1,0 @@
-.root-layout {
-  
-  .footer {
-    margin-top: 20px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 40px;
-    border-top: 1px solid #eaeaea;
-
-    .top {
-      display: flex;
-      gap: 30px;
-      margin-top: 20px;
-
-    }
-
-    .bottom {
-      padding-bottom: 30px;
-    }
-  }
-}
