+++ components/book/BooksList/BookList.module.scss
@@ -24,7 +24,6 @@
   }
 }
 
-
 /* Extra small devices (phones, 600px and down) */
 @media only screen and (max-width: 600px) {
   .grid {
@@ -46,7 +45,7 @@
   }
 }
 
-@media only screen and (min-width: 992px)  {
+@media only screen and (min-width: 992px) {
   .grid {
     grid-template-columns: repeat(3, 1fr);
   }
