+++ components/common/Footer/Footer.tsx
@@ -1,7 +1,16 @@
-import React from 'react'
+import styles from './Footer.module.scss'
 
 const Footer = () => {
-  return <footer>This is a footer</footer>
+  return (
+    <footer className={styles.footer}>
+      <div className={styles['top']}>
+        <div>About Page</div>
+        <div>Contact</div>
+      </div>
+
+      <div className={styles['bottom']}>Copyright ...</div>
+    </footer>
+  )
 }
 
 export default Footer
